Refactor to try out traits
Note that it doesn't actually work in this state, because
it's no longer deserializing the top level json key away.

{
    "Global Quote": {
        "01. symbol": "GOOG",
        "02. open": "1104.2500",
        "03. high": "1106.0300",
        "04. low": "1048.7625",
        "05. price": "1050.7100",
        "06. volume": "1625564",
        "07. latest trading day": "2018-10-24",
        "08. previous close": "1103.6900",
        "09. change": "-52.9800",
        "10. change percent": "-4.8003%"
    }
}

I.e., in that, the "Global Quote" is no longer being handled. Next
order of business should probably be learning more about serde
serialization and deserialization to customize that for a given
type of response.